requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==3.2.4
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
autopep8==2.3.1
babel==2.16.0
bandit==1.7.9
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
cryptography==43.0.1
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==4.4.2
deepgram-sdk==3.7.3
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.8
distlib==0.3.8
distro==1.9.0
docutils==0.21.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.0
flake8==7.1.1
fqdn==1.5.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
imageio==2.36.0
imageio-ffmpeg==0.5.1
importlib_metadata==8.4.0
ipykernel==6.29.5
ipython==8.27.0
isoduration==20.11.0
isort==5.13.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.4
jiter==0.6.1
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keyring==25.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.5.0
moviepy==1.0.3
multidict==6.1.0
mypy==1.11.2
mypy-extensions==1.0.0
narwhals==1.9.4
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nh3==0.2.18
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.1.2
openai==1.51.2
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pbr==6.1.0
pep8==1.7.1
pexpect==4.9.0
pillow==10.4.0
pipenv==2024.0.1
pkginfo==1.10.0
platformdirs==4.3.2
proglog==0.1.10
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.2.0
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydocstyle==6.3.0
pyflakes==3.2.0
Pygments==2.18.0
pylama==8.4.1
pylint==3.2.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytoolconfig==1.3.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
readme_renderer==44.0
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.8.0
rope==1.13.0
rpds-py==0.20.0
SecretStorage==3.3.3
Send2Trash==1.8.3
setuptools==74.1.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
stack-data==0.6.3
stevedore==5.3.0
streamlit==1.39.0
tenacity==9.0.0
terminado==0.18.1
tinycss2==1.3.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
twine==5.1.1
types-python-dateutil==2.9.0.20240906
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.2
virtualenv==20.26.4
watchdog==5.0.3
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
wheel==0.44.0
yarl==1.15.4
zipp==3.20.1
